requirements.txt
alembic==1.17.2
et-xmlfile==2.0.0
greenlet==3.2.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
psycopg==3.2.9
psycopg-binary==3.2.9
pyside6==6.9.0
pyside6-addons==6.9.0
pyside6-essentials==6.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
shiboken6==6.9.0
six==1.17.0
sqlalchemy==2.0.41
typing-extensions==4.13.2
tzdata==2025.2
